Update README.md
Data collection: The first step is to collect data that contains text that you want to analyze. This data can come from a variety of sources, such as social media, customer reviews, or surveys.
Data preprocessing: Once the data is collected, it needs to be preprocessed. This involves cleaning the data and removing any irrelevant or noisy information.
Feature extraction: The next step is to extract features from the text. Features are the words or phrases that will be used to train the sentiment analysis model.
Model training: The features are then used to train a sentiment analysis model. This model learns to associate certain features with positive or negative sentiment.
Model evaluation: Once the model is trained, it needs to be evaluated. This is done by testing the model on a set of data that was not used to train the model.
Model deployment: Once the model is evaluated and found to be accurate, it can be deployed. This means that the model can be used to predict the sentiment of new text.